Move PemToKeyPairTask to actual Task
