+++ app/src/main/java/com/amaze/filemanager/exceptions/SshKeyInvalidPassphrase.kt
@@ -1,0 +1,6 @@
+package com.amaze.filemanager.exceptions
+
+import java.io.IOException
+
+class SshKeyInvalidPassphrase : IOException() {
+}
